+++ src/App.tsx
@@ -3,31 +3,25 @@ import { Navbar } from './components/Navbar';
 import './App.scss';
 import { Navigate, Route, Routes } from 'react-router-dom';
 import { PeoplePage } from './components/PeoplePage';
-import { PeopleContextProvider } from './context/PeopleContext';
 
 export const App = () => {
   return (
-    <PeopleContextProvider>
-      <div data-cy="app">
-        <Navbar />
+    <div data-cy="app">
+      <Navbar />
 
-        <main className="section">
-          <div className="container">
-            <Routes>
-              <Route path="/home" element={<Navigate to="/" replace />} />
-              <Route
-                path="/"
-                element={<h1 className="title">People Page</h1>}
-              />
-              <Route path="/people/:personName?" element={<PeoplePage />} />
-              <Route
-                path="*"
-                element={<h1 className="title">Page not found</h1>}
-              />
-            </Routes>
-          </div>
-        </main>
-      </div>
-    </PeopleContextProvider>
+      <main className="section">
+        <div className="container">
+          <Routes>
+            <Route path="/home" element={<Navigate to="/" replace />} />
+            <Route path="/" element={<h1 className="title">People Page</h1>} />
+            <Route path="/people/:personName?" element={<PeoplePage />} />
+            <Route
+              path="*"
+              element={<h1 className="title">Page not found</h1>}
+            />
+          </Routes>
+        </div>
+      </main>
+    </div>
   );
 };
