--- src/utils/noop.ts
@@ -1,1 +1,0 @@
-export const noop = () => {};
